pyproject.toml
[project]
name = "vault-aegis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pyjwt>=2.0.0",
    "redis>=4.0.0",
    "pyyaml>=6.0",
    "pydantic>=1.8.0",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]
